dmatrix: Fix routines that cause bug in example/colon
example/colon fails to resize the matrix in the
following line:
	A(colon(),6) = ones(5,1);

This issue is caused by one routine that does not check
the number of columns causing an invalid memory access.
This patch fixes the routine:
	DMatrix& DMatrix::operator() (const DMatrix& RowIndx, long col )

However, this issue is also reproduced by the following
routine:
	DMatrix& DMatrix::operator() (long row, const DMatrix& ColIndx )

These two routines are modified by this patch, testing
the number of cols/rows and resizing the matrix before
accessing their elements.

Signed-off-by: Flavio Santes <flavio.santes@1byt3.com>